Split FlyoutBase and PopupFlyoutBase
